requirements.txt
numpy
openpyxl
pandas
scikit-learn
python-dotenv
streamlit
python-Levenshtein
